requirements.txt
# fix some problems with the queue
runpod~=1.7.12
websocket-client
requests

gradio
Pillow
